sighthouse-frontend/pyproject.toml
[project]
name = "sighthouse.frontend"
description = "Frontend sighthouse functionality"
authors = [
  {name="Benoît FORGETTE aka MadSquirrel"},
  {name="Sami BABIGEON aka Fenrisfulsur"}
]
license = "MIT"
version = "1.0.0"
dependencies = [
  "celery>=5.5.3", # frontend/pipeline
  "redis>=6.3.0", # nowhere??
  "Flask-Login>=0.6.3",
  "numpy==2.4.2",
  "Werkzeug>=3.1.3",
  "cxxfilt",
  "sighthouse-core>=1.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.txt", "**/*.java"]

[tool.black]
line-length = 100

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
